chore: rename checkEthFlowOrderExists
